SuitRecharger: the main logic was separated into Workable and StateMachineComponent.
this way it will be easier to add multiple Chores.
