+++ src/SuitRecharger/SuitRecharger.csproj
@@ -30,6 +30,7 @@
       <Link>Utils.cs</Link>
     </Compile>
     <Compile Include="AlwaysFunctionalConduitDispenser.cs" />
+    <Compile Include="SuitRechargerWorkable.cs" />
     <Compile Include="STRINGS.cs" />
     <Compile Include="SuitRecharger.cs" />
     <Compile Include="SuitRechargerPatches.cs" />
